% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{address}
\alias{address}
\title{address}
\usage{
address(latlong = NULL)
}
\arguments{
\item{latlong}{A character string.}
}
\value{
A dataframe cointaining full address information
}
\description{
The 'address' function is a function that convert the coordinates to address.
}
\examples{
address(latlong = "58.39701, 15.57415")

}
