% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_lookup.R
\name{address_lookup}
\alias{address_lookup}
\title{Address finder}
\usage{
address_lookup(latlong = NULL)
}
\arguments{
\item{latlong}{A character string that includes the latitude and the longitude.}
}
\value{
A dataframe cointaining Full address names as well as the coordinates
}
\description{
The 'address_lookup' function is a function that provides the user the address of a given
a set of coordinates The funtion uses the google geocode API to retrieve the result.
}
\examples{
address_lookup(latlong = "58.39701, 15.57415")

}
