% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\docType{package}
\name{boxgooglemap}
\alias{boxgooglemap}
\alias{boxgooglemap-package}
\title{boxgooglemap: Transfer between address and coordinates use the google geocode API.}
\description{
Transfer between address and coordinates use the google geocode API.
}
\references{
\url{https://developers.google.com/maps/documentation/geocoding/start}
}
\author{
Maintainer: Boxi Zhang bossyzhang@gmail.com
}
