% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{coord}
\alias{coord}
\title{Coordinate finder}
\usage{
coord(address = NULL)
}
\arguments{
\item{address}{A character string that includes address components}
}
\value{
A dataframe cointaining full address names as well as the coordinates
}
\description{
The 'coord' function is a function that convert the address to the coordinates.
}
\examples{
coord(address = "Chang Sha, China")

}
