% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_lookup.R
\name{coord_lookup}
\alias{coord_lookup}
\title{Coordinate finder}
\usage{
coord_lookup(address = NULL)
}
\arguments{
\item{address}{A character string that includes address components}
}
\value{
A dataframe cointaining Full address names as well as the coordinates
}
\description{
The 'coord_lookup' function is a function that provides the user the coordinates of a given
address. The funtion uses the google geocode API to retrieve the result.
}
\examples{
coord_lookup(address = "Mäster Mattias väg, Linköping")

}
